add switch and interactive cli mode

+++ collage_cli.py
@@ -14,10 +14,23 @@
 #     You should have received a copy of the GNU General Public License
 #     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from collage.collage import *
 import glob
 import os
+import cv2
+from prompt_toolkit.validation import Validator
+from prompt_toolkit import prompt
 import argparse
+
+from collage.collage import ImageList, Collage, CollageLayout
+
+def is_number(text):
+    return text.isdigit()
+
+validator = Validator.from_callable(
+    is_number,
+    error_message='This input contains non-numeric characters',
+    move_cursor_to_end=True)
+
 
 if __name__ == "__main__":
     ''' Example usage of collage
@@ -56,9 +69,47 @@ if __name__ == "__main__":
 
     target_image = my_collage.render_image()
 
+    cv2.namedWindow('image')
     cv2.imshow("image", target_image)
     cv2.imwrite("collage.jpg", target_image)
-    cv2.waitKey(0)
+    cv2.waitKey(10)
 
-    cv2.destroyAllWindows()
+    exit_program = False
+    while not exit_program:
+        text = prompt('> ')
+
+        print('You said: %s' % text)
+        if text == 'exit':
+            exit_program = True
+            cv2.destroyAllWindows()
+
+        if text in ['h', 'help']:
+            print('available commands:')
+            print('exit: exit the program')
+            print('h, help: print this help')
+            print('i, indices: toggle display of image indices in collage')
+            print('save: save the collage as collage.jpg')
+            print('s, switch: switch the images at the given indices')
+
+        if text == 'save':
+            save_filename = 'collage.jpg'
+            cv2.imwrite(save_filename, target_image)
+            print(f'Wrote collage to {save_filename}')
+
+        if text in ['s', 'switch']:
+            image_1 = int(prompt('First image index: ', validator=validator)) - 1
+            image_2 = int(prompt('Second image index: ', validator=validator)) - 1
+            print(f'switching images {image_1+1} and {image_2+1}')
+            target_image = my_collage.switch(image_1, image_2)
+            cv2.imshow("image", target_image)
+            cv2.waitKey(10)
+
+        if text in ['i', 'indices']:
+            if my_collage.show_indices:
+                my_collage.show_indices = False
+            else:
+                my_collage.show_indices = True
+            target_image = my_collage.render_image()
+            cv2.imshow("image", target_image)
+            cv2.waitKey(10)
